refactor: move internal build logic to cli-build package
