+++ packages/@sanity/cli-build/src/_exports/_internal/build.ts
@@ -1,15 +1,15 @@
-export {buildDebug} from '../../actions/build/buildDebug.js'
+export {buildApp} from '../../actions/build/buildApp.js'
 export {buildStaticFiles} from '../../actions/build/buildStaticFiles.js'
+export {buildStudio} from '../../actions/build/buildStudio.js'
 export {checkRequiredDependencies} from '../../actions/build/checkRequiredDependencies.js'
 export {checkStudioDependencyVersions} from '../../actions/build/checkStudioDependencyVersions.js'
-export {
-  getAutoUpdatesCssUrls,
-  getAutoUpdatesImportMap,
-  getModuleUrl,
-} from '../../actions/build/getAutoUpdatesImportMap.js'
+export {getModuleUrl} from '../../actions/build/getAutoUpdatesImportMap.js'
 export {extendViteConfigWithUserConfig, getViteConfig} from '../../actions/build/getViteConfig.js'
 export {resolveVendorBuildConfig} from '../../actions/build/resolveVendorBuildConfig.js'
 export {writeSanityRuntime} from '../../actions/build/writeSanityRuntime.js'
 export {SANITY_CACHE_DIR} from '../../constants.js'
-export {AppBuildTrace, StudioBuildTrace} from '../../telemetry/build.telemetry.js'
-export {formatModuleSizes, sortModulesBySize} from '../../util/moduleFormatUtils.js'
+export type {
+  CompareDependencyVersions,
+  CompareDependencyVersionsResult,
+  UnresolvedPrerelease,
+} from '../../util/compareDependencyVersions.js'
