+++ packages/@sanity/cli-build/src/util/compareDependencyVersions.d.ts
@@ -1,0 +1,15 @@
+export interface CompareDependencyVersions {
+  installed: string
+  pkg: string
+  remote: string
+}
+
+export interface UnresolvedPrerelease {
+  pkg: string
+  version: string
+}
+
+export interface CompareDependencyVersionsResult {
+  mismatched: Array<CompareDependencyVersions>
+  unresolvedPrerelease: Array<UnresolvedPrerelease>
+}
